+++ sapling/std/classes/__init__.py
@@ -3,8 +3,10 @@ from .unicode import Unicode
 from .runtime import Runtime
 from .threads import Threads
 from .parsers import Parsers
+from .logger import Logger
 from .system import System
 from .math import Math
+from .bits import Bits
 from .py import Py
 
 
@@ -15,5 +17,7 @@ public_classes = {
     'Py': Class.from_py_cls(Py, -1, -1),
     'Runtime': Class.from_py_cls(Runtime, -1, -1),
     'Parsers': Class.from_py_cls(Parsers, -1, -1),
-    'Unicode': Class.from_py_cls(Unicode, -1, -1)
+    'Unicode': Class.from_py_cls(Unicode, -1, -1),
+    'Bits': Class.from_py_cls(Bits, -1, -1),
+    'Logger': Class.from_py_cls(Logger, -1, -1),
 }