+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 ---
 
+### v0.1.3
+- Importing From
+- Standard library improvement
+- && and || operators
+- Bug fixes
 ### v0.1.2.2
 - Starts with a main function (if a main function exists)
 - Bug fixes
@@ -9,10 +14,10 @@
 ### v0.1.2
 - Struct/Attr Function Definitions
 - Repeat Statement
-- Standard library improvement/enhancement
+- Standard library improvement
 - Local file imports
 - Bug fixes
-- Objects performance improvements
+- Objects performance improvements (using \_\_slots\_\_)
 ### v0.1.1
 - Augmented assignment operators
 - Indexing